+++ packages/wpcom.js/test/test.wpcom.site.js
@@ -226,6 +226,144 @@ describe('wpcom.site', function () {
       });
     });
 
+    describe('wpcom.site.statsComments', function () {
+      it('should request comments data', function (done) {
+        site.statsComments(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.ok(data.authors instanceof Array);
+          assert.ok(data.posts instanceof Array);
+          assert.equal('number', typeof data.monthly_comments);
+          assert.equal('number', typeof data.total_comments);
+          assert.equal('string', typeof data.most_active_day);
+          assert.equal('string', typeof data.most_active_time);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsCommentFollowers', function () {
+      it('should request comment follower data', function (done) {
+        site.statsCommentFollowers(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.ok(data.posts instanceof Array);
+          assert.equal('number', typeof data.page);
+          assert.equal('number', typeof data.pages);
+          assert.equal('number', typeof data.total);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsFollowers', function () {
+      it('should request follower data', function (done) {
+        site.statsFollowers(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.ok(data.subscribers instanceof Array);
+          assert.equal('number', typeof data.page);
+          assert.equal('number', typeof data.pages);
+          assert.equal('number', typeof data.total);
+          assert.equal('number', typeof data.total_email);
+          assert.equal('number', typeof data.total_wpcom);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsPublicize', function () {
+      it('should request publicize data', function (done) {
+        site.statsPublicize(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.ok(data.services instanceof Array);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsStreak', function () {
+      it('should request streak data', function (done) {
+        site.statsStreak(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.ok(data.streak instanceof Object);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsSummary', function () {
+      it('should request summary data', function (done) {
+        site.statsSummary(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.equal('string', typeof data.period);
+          assert.equal('number', typeof data.likes);
+          assert.equal('number', typeof data.views);
+          assert.equal('number', typeof data.visitors);
+          assert.equal('number', typeof data.comments);
+          assert.equal('number', typeof data.followers);
+          assert.equal('number', typeof data.reblogs);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsTags', function () {
+      it('should request tag data', function (done) {
+        site.statsTags(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.ok(data.tags instanceof Array);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsTopAuthors', function () {
+      it('should request author data', function (done) {
+        site.statsTopAuthors(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.equal('string', typeof data.period);
+          assert.ok(data.days instanceof Object);
+
+          done();
+        });
+      });
+    });
+
+    describe('wpcom.site.statsVideoPlays', function () {
+      it('should request video play data', function (done) {
+        site.statsVideoPlays(function (err, data) {
+          if (err) throw err;
+
+          assert.equal('string', typeof Date(data.day));
+          assert.equal('string', typeof data.period);
+          assert.ok(data.days instanceof Object);
+
+          done();
+        });
+      });
+    });
+
     describe('wpcom.site.statsVisits', function () {
       it('should request visits stats', function (done) {
         site.statsVisits(function (err, data) {
@@ -246,16 +384,12 @@ describe('wpcom.site', function () {
 
     describe('wpcom.site.statsReferrers', function () {
       it('should request referrers stats', function (done) {
-        site.statsReferrers({ apiVersion: '1' }, function (err, data) {
+        site.statsReferrers( function (err, data) {
           if (err) throw err;
 
           assert.equal('string', typeof Date(data.date));
-
-          assert.equal('number', typeof data.days);
-          assert.equal('number', typeof data.total);
-
-          assert.equal('object', typeof data.referrers);
-          assert.ok(data.referrers instanceof Array);
+          assert.equal('object', typeof data.days);
+          assert.equal('string', typeof data.period);
 
           done();
         });
@@ -264,12 +398,12 @@ describe('wpcom.site', function () {
 
     describe('wpcom.site.statsTopPosts', function () {
       it('should request top posts stats', function (done) {
-        site.statsTopPosts({ apiVersion: '1' }, function (err, data) {
+        site.statsTopPosts( function (err, data) {
           if (err) throw err;
 
           assert.equal('string', typeof Date(data.date));
-          assert.equal('object', typeof data['top-posts']);
-          assert.ok(data['top-posts'] instanceof Array);
+          assert.equal('object', typeof data.days);
+          assert.equal('string', typeof data.period);
 
           done();
         });
@@ -278,12 +412,12 @@ describe('wpcom.site', function () {
 
     describe('wpcom.site.statsCountryViews', function () {
       it('should request country views stats', function (done) {
-        site.statsCountryViews({ apiVersion: '1' }, function (err, data) {
+        site.statsCountryViews( function (err, data) {
           if (err) throw err;
 
           assert.equal('string', typeof Date(data.date));
-          assert.equal('object', typeof data['country-views']);
-          assert.ok(data['country-views'] instanceof Array);
+          assert.equal('object', typeof data.days);
+          assert.ok(data['country-info'] instanceof Array);
 
           done();
         });
@@ -292,16 +426,12 @@ describe('wpcom.site', function () {
 
     describe('wpcom.site.statsClicks', function () {
       it('should request clicks stats', function (done) {
-        site.statsClicks({ apiVersion: '1' }, function (err, data) {
+        site.statsClicks( function (err, data) {
           if (err) throw err;
 
           assert.equal('string', typeof Date(data.date));
-
-          assert.equal('number', typeof data.days);
-          assert.equal('number', typeof data.total);
-
-          assert.equal('object', typeof data.clicks);
-          assert.ok(data.clicks instanceof Array);
+          assert.equal('object', typeof data.days);
+          assert.equal('string', typeof data.period);
 
           done();
         });
@@ -310,15 +440,12 @@ describe('wpcom.site', function () {
 
     describe('wpcom.site.statsSearchTerms', function () {
       it('should request search terms stats', function (done) {
-        site.statsSearchTerms({ apiVersion: '1' }, function (err, data) {
+        site.statsSearchTerms( function (err, data) {
           if (err) throw err;
 
           assert.equal('string', typeof Date(data.date));
-
-          assert.equal('number', typeof data.days);
-
-          assert.equal('object', typeof data['search-terms']);
-          assert.ok(data['search-terms'] instanceof Array);
+          assert.equal('object', typeof data.days);
+          assert.equal('string', typeof data.period);
 
           done();
         });
@@ -389,6 +516,55 @@ describe('wpcom.site', function () {
         assert.ok(data);
         assert.ok(data.media instanceof Array);
         assert.equal(fixture.media.urls.length, data.media.length);
+        done();
+      });
+    });
+  });
+
+  describe('wpcom.site.statsReferrersSpamNew', function() {
+    var d = new Date();
+    var domain = ( d.getTime() / 1000 ) + 'wordpress.com';
+    it('should mark a domain as spam', function (done) {
+      site.statsReferrersSpamNew( domain, function(err, data) {
+        if (err) throw err;
+
+        assert.ok(data);
+        done();
+      });
+    });
+  });
+
+  describe('wpcom.site.statsReferrersSpamDelete', function() {
+    var d = new Date();
+    var domain = ( d.getTime() / 1000 ) + 'wordpress.com';
+    it('should remove a domain from spam refferer list', function (done) {
+      site.statsReferrersSpamDelete( domain, function(err, data) {
+        if (err) throw err;
+
+        assert.ok(data);
+        done();
+      });
+    });
+  });
+
+  describe('wpcom.site.statsPostViews', function() {
+    it('should request post stat details', function (done) {
+      site.statsPostViews( testing_post.ID, function(err, data) {
+        if (err) throw err;
+
+        assert.ok(data);
+        assert.equal('string', typeof Date(data.date));
+        assert.equal('number', typeof data.views);
+        assert.equal('number', typeof data.highest_month);
+        assert.equal('number', typeof data.highest_day_average);
+        assert.equal('number', typeof data.highest_week_average);
+        assert.ok(data.post instanceof Object);
+        assert.ok(data.years instanceof Object);
+        assert.ok(data.weeks instanceof Object);
+        assert.ok(data.fields instanceof Array);
+        assert.ok(data.data instanceof Array);
+        assert.ok(data.averages instanceof Object);
+
         done();
       });
     });